+++ sdk/core/azure_core/src/http/policies/instrumentation/request_instrumentation.rs
@@ -42,8 +42,7 @@ impl RequestInstrumentationPolicy {
     }
 }
 
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[async_trait::async_trait]
 impl Policy for RequestInstrumentationPolicy {
     async fn send(
         &self,