+++ sdk/core/azure_core_test_macros/src/test.rs
@@ -97,7 +97,6 @@ pub fn parse_test(attr: TokenStream, item: TokenStream) -> Result<TokenStream> {
     outer_sig.inputs.clear();
 
     Ok(quote! {
-        #[cfg(not(target_arch = "wasm32"))]
         #test_attr
         #(#attrs)*
         #vis #outer_sig {