+++ sdk/cosmos/azure_data_cosmos/src/pipeline/authorization_policy.rs
@@ -61,8 +61,7 @@ impl AuthorizationPolicy {
     }
 }
 
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[async_trait::async_trait]
 impl Policy for AuthorizationPolicy {
     async fn send(
         &self,
@@ -157,8 +156,7 @@ mod tests {
     #[derive(Debug)]
     struct TestTokenCredential(String);
 
-    #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-    #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+    #[async_trait::async_trait]
     impl TokenCredential for TestTokenCredential {
         async fn get_token(
             &self,
@@ -267,8 +265,7 @@ mod tests {
             captured_scopes: Arc<Mutex<Vec<Vec<String>>>>,
         }
 
-        #[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-        #[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+        #[async_trait::async_trait]
         impl TokenCredential for ScopeCapturingCredential {
             async fn get_token(
                 &self,