+++ sdk/identity/azure_identity/src/client_secret_credential.rs
@@ -115,8 +115,7 @@ impl ClientSecretCredential {
     }
 }
 
-#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
-#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[async_trait::async_trait]
 impl TokenCredential for ClientSecretCredential {
     async fn get_token(
         &self,