+++ sdk/keyvault/azure_security_keyvault_secrets/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Breaking Changes
 
+- Support for `wasm32-unknown-unknown` has been removed ([#3377](https://github.com/Azure/azure-sdk-for-rust/issues/3377))
+
 ### Bugs Fixed
 
 ### Other Changes