+++ sdk/storage/azure_storage_queue/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Breaking Changes
 
 - Changed `QueueClient::set_metadata()` `metadata` parameter from owned `HashMap<String, String>` to `&HashMap<String, String>`.
+- Support for `wasm32-unknown-unknown` has been removed ([#3377](https://github.com/Azure/azure-sdk-for-rust/issues/3377))
 
 ### Bugs Fixed